// src/api/ocgcore/idl/ocgcore.ts
export namespace ygopro {
// ...
    export namespace StocTypeChange {
        export enum SelfType {
            UNKNOWN = 0,
            PLAYER1 = 1,
            PLAYER2 = 2,
            PLAYER3 = 3,
            PLAYER4 = 4,
            PLAYER5 = 5,
            PLAYER6 = 6,
            OBSERVER = 100
        }
    }
// ...
}